LS-21495: Enable warning or critical threshold to be null
I discovered that if you call something like d.Get("thresholds")
and it returns a map of things, all those things will have the
primitive default values. This will cause us to set optional
floats to zero when they should be nil.

Instead you need to call d.getOK() on the exact object in order
to find out if it exists.