LS-42929 add support for composite alerts (#144)
Composite alerts! In terraform!

Composite alerts allow a user to create an alert on multiple signals with their own thresholds. (Query A is above Threshold A && Query B is below Threshold B && ...)

A lightstep_alert resource now no longer allows an expression or query block be provided if composite_alert is provided. Composite alert queries must be written in UQL.