Turn off acceptance tests for dependabot PRs (#162)
PRs triggered by dependabot are not given access to the secrets configured for Github actions. This causes the acceptance tests to fail because there is no Lightstep api key. The reasoning for dependabot working this way is to prevent dependencies from exfiltrating those secrets or the other data that they have access to. (See dependabot/dependabot-core#3253 (comment))

This means that we really should have a human look at the dependency updates before blindly running the acceptance tests. I think a reasonable solution to this is to disable the acceptance tests on dependabot PRs and add acceptance tests to the release workflow so that we know they always pass before releasing a new version.